requirements.txt
tensorflow~=2.18.0
numpy~=2.0.2
Flask~=3.1.0
pillow~=11.0.0
matplotlib~=3.9.3